+++ prothos/runtime/mcconf.module
@@ -1,4 +1,4 @@
 # -*- mode:toml; -*-
 [module.runtime-prothos]
-    incfiles = [ "runtime/Task.hh", "runtime/DAG.hh", "utils/hlmsDeque.hh","utils/shim.hh", "runtime/Worker.hh", "runtime/Thread.hh", "utils/FifoQueue.hh", "runtime/FlowGraph.hh" ]
+    incfiles = [ "runtime/Task.hh", "runtime/DAG.hh", "utils/hlmsDeque.hh","utils/shim.hh", "runtime/Worker.hh", "runtime/Thread.hh", "utils/FifoQueue.hh", "runtime/FlowGraph.hh", "runtime/FlowGraph_impl.hh"]
     appfiles = [ "runtime/Task.cc", "runtime/Thread.cc"]
